refactor(tenant): split server.go into domain endpoint files
Extract RPC handlers from server.go (724 lines) into focused files:
- grpc_tenant_endpoints.go: InitiateTenant, RetrieveTenant, GetBySlug, UpdateTenantStatus, ListTenants
- grpc_provisioning_endpoints.go: ReconcileMigrations, GetTenantProvisioningStatus, createProvisioningStatusRecords
- mappers.go: toProto, toProtoStatus, toDomainStatus, toProtoServiceStatus

server.go retains only the Service struct, constructor, ErrUnknownStatus, and provisioningHintFromStatus.
Pure code motion - no behavioral changes.
